add the share popup for event line
