+++ src/components/ComponentPages/eventLine/TimelineSlider/index.tsx
@@ -1,5 +1,14 @@
 import { useEffect, useRef, useState } from "react";
-import { Slider, Popover, Typography, Button } from "antd";
+import {
+  Slider,
+  Popover,
+  Typography,
+  Button,
+  Radio,
+  Space,
+  Tooltip,
+  RadioChangeEvent,
+} from "antd";
 import {
   CaretRightOutlined,
   PauseOutlined,
@@ -9,6 +18,18 @@ import {
 } from "@ant-design/icons";
 import styles from "./timeBarControl.module.scss";
 import { useRouter } from "next/router";
+import {
+  FacebookIcon,
+  FacebookShareButton,
+  LinkedinIcon,
+  LinkedinShareButton,
+  TwitterIcon,
+  TwitterShareButton,
+} from "next-share";
+import { isServer } from "src/utils/generalUtility";
+import Paragraph from "antd/lib/typography/Paragraph";
+import CopyLinkIcon from "../../../../assets/image/copy-link.png";
+import CheckIcon from "../../../../assets/image/check.png";
 
 const { Title } = Typography;
 
@@ -37,17 +58,97 @@ function TimelineSlider({
   handleForwardOrBackord,
   isPlaying,
   setIsPlaying,
-  eventId,
-  setEventId,
-  value,
-  setValue,
-}: any) {
+}: // eventId,
+// setEventId,
+// value,
+// setValue,
+any) {
   const router = useRouter();
   const didMount = useRef(false);
-
   const [intervalId, setIntervalId] = useState(null);
-
   const [speedBar, setSpeedBar] = useState(false);
+  const [URL, setURL] = useState("");
+  const [eventId, setEventId] = useState(0);
+  const [value, setValue] = useState(1);
+
+  const updateEventId = (url, newValue) => {
+    let urlParts = url.split("?");
+    let baseUrl = urlParts[0];
+    let params = urlParts[1].split("&");
+
+    for (let i = 0; i < params.length; i++) {
+      let paramParts = params[i].split("=");
+      if (paramParts[0] === "eventId") {
+        paramParts[1] = newValue;
+        params[i] = paramParts.join("=");
+        break; // Found eventId, no need to continue looping
+      }
+    }
+
+    return baseUrl + "?" + params.join("&");
+  };
+
+  const optionChange = (e: RadioChangeEvent) => {
+    setValue(e.target.value);
+
+    if (e.target.value == 1) {
+      if (router.asPath.includes("?")) {
+        setURL(!isServer() && window?.location?.href?.split("?")[0]);
+      } else {
+        setURL(!isServer() && window?.location?.href);
+      }
+    } else if (e.target.value == 2) {
+      if (router.asPath.includes("?")) {
+        if (router.asPath.includes("eventId")) {
+          setURL(updateEventId(window?.location?.href, eventId));
+        } else {
+          setURL(window?.location?.href + `&eventId=${eventId}`);
+        }
+      } else {
+        setURL(window?.location?.href + `?eventId=${eventId}`);
+      }
+    }
+  };
+
+  const copyHandler = () => {
+    if (value == 1) {
+      if (router.asPath.includes("?")) {
+        setURL(!isServer() && window?.location?.href?.split("?")[0]);
+      } else {
+        setURL(!isServer() && window?.location?.href);
+      }
+    } else if (value == 2) {
+      if (router.asPath.includes("?")) {
+        if (router.asPath.includes("eventId")) {
+          setURL(updateEventId(window?.location?.href, eventId));
+        } else {
+          setURL(window?.location?.href + `&eventId=${eventId}`);
+        }
+      } else {
+        setURL(window?.location?.href + `?eventId=${eventId}`);
+      }
+    }
+  };
+
+  useEffect(() => {
+    if (value == 1) {
+      if (router.asPath.includes("?")) {
+        setURL(!isServer() && window?.location?.href?.split("?")[0]);
+      } else {
+        setURL(!isServer() && window?.location?.href);
+      }
+    } else if (value == 2) {
+      if (router.asPath.includes("?")) {
+        if (router.asPath.includes("eventId")) {
+          setURL(updateEventId(window?.location?.href, eventId));
+        } else {
+          setURL(window?.location?.href + `&eventId=${eventId}`);
+        }
+      } else {
+        setURL(window?.location?.href + `?eventId=${eventId}`);
+      }
+    }
+  }, [eventId, value]);
 
   useEffect(() => {
     if (!didMount.current) {
@@ -158,7 +259,7 @@ function TimelineSlider({
     setSpeedBar(false);
   };
 
-  const content = (
+  const controller_content = (
     <div className="speed-controller">
       <Title level={4}>Playback speed</Title>
       <Slider
@@ -290,6 +391,64 @@ function TimelineSlider({
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [iteration, mockData]);
 
+  const content = (
+    <div className={styles.popoverWrapper}>
+      <Title level={5}>Share</Title>
+      <Radio.Group onChange={optionChange} value={value}>
+        <Space direction="vertical">
+          <Radio value={1}>Eventline URL</Radio>
+          <Radio value={2}>Current Event URL</Radio>
+        </Space>
+      </Radio.Group>
+      <Space align={"center"} className={styles.shareLink}>
+        <Title level={5}> Share Link To: </Title>
+        <Space>
+          <FacebookShareButton
+            url={URL}
+            hashtag={`#${!isServer() && window?.location?.hostname}`}
+          >
+            <Tooltip title="Share On Facebook">
+              <FacebookIcon size={27} round />
+            </Tooltip>
+          </FacebookShareButton>
+
+          <TwitterShareButton url={URL}>
+            <Tooltip title="Share On Twitter">
+              <TwitterIcon size={27} round />
+            </Tooltip>
+          </TwitterShareButton>
+
+          <LinkedinShareButton url={URL}>
+            <Tooltip title="Share On Linkedin">
+              <LinkedinIcon size={27} round />
+            </Tooltip>
+          </LinkedinShareButton>
+          <>
+            <Paragraph
+              className={styles.typographyLink}
+              copyable={{
+                text: URL,
+                icon: [
+                  <img
+                    src={CopyLinkIcon.src}
+                    key="1"
+                    style={{ verticalAlign: "baseline" }}
+                    onClick={() => copyHandler()}
+                  />,
+                  <img
+                    src={CheckIcon.src}
+                    key="2"
+                    style={{ verticalAlign: "baseline" }}
+                  />,
+                ],
+              }}
+            ></Paragraph>
+          </>
+        </Space>
+      </Space>
+    </div>
+  );
+
   return (
     <>
       <div className="player-wrapper">
@@ -338,7 +497,7 @@ function TimelineSlider({
             data-testid="forward-button"
           ></i>
           <Popover
-            content={content}
+            content={controller_content}
             title={false}
             trigger="hover"
             open={speedBar}
@@ -357,12 +516,14 @@ function TimelineSlider({
           </Popover>
         </div>
         <div className="share-wrapper">
-          <Button
-            type="link"
-            className="text-canBlack"
-            icon={<ShareAltOutlined />}
-            size="small"
-          />
+          <Popover content={content} trigger="click" placement="leftBottom">
+            <Button
+              type="link"
+              className="text-canBlack"
+              icon={<ShareAltOutlined />}
+              size="small"
+            />
+          </Popover>
         </div>
       </div>
       {mockData && (
